requirements.txt
# Kafka dependencies
kafka-python==2.0.2

# Database dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
pandas==2.1.4
numpy==1.24.4

# Dashboard
streamlit==1.29.0
plotly==5.17.0

# Utilities
python-dotenv==1.0.0